requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
Django==5.0.6
django-redis==5.4.0
djangorestframework==3.15.1
mysqlclient==2.2.4
pillow==10.3.0
redis==5.0.4
sqlparse==0.5.0
typing-extensions==4.12.0
tzdata==2024.1
